added dialog boxes.

+++ src/main/java/controller/deworming/IndexController.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 import org.apache.logging.log4j.LogManager;
@@ -21,11 +22,15 @@ import javafx.beans.property.ReadOnlyObjectWrapper;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+import javafx.event.Event;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeTableColumn;
 import javafx.stage.Modality;
@@ -35,9 +40,6 @@ import model.Desparasitaciones;
 import model.Pacientes;
 
 public class IndexController {
-    protected static final Logger log = (Logger) LogManager.getLogger(IndexController.class);
-    static DesparasitacionesHome dao = new DesparasitacionesHome();
-
     @FXML
     private ResourceBundle resources;
 
@@ -55,6 +57,10 @@ public class IndexController {
 
     @FXML
     private JFXTreeTableView<Desparasitaciones> indexD;
+
+    protected static final Logger log = (Logger) LogManager.getLogger(IndexController.class);
+
+    private static DesparasitacionesHome dao = new DesparasitacionesHome();
 
     private Desparasitaciones desparacitaciones;
 
@@ -114,41 +120,78 @@ public class IndexController {
         });
 
         btnEdit.setOnAction((event) -> {
-            Parent rootNode;
-            Stage stage = new Stage();
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/fxml/deworming/modalDialog.fxml"));
-            Window node = ((Node) event.getSource()).getScene().getWindow();
-            try {
-                rootNode = (Parent) fxmlLoader.load();
-                ModalDialogController mdc = fxmlLoader.getController();
-                mdc.setObject(desparacitaciones);
-                stage.setScene(new Scene(rootNode));
-                stage.setTitle("Desparasitaciones");
-                stage.initModality(Modality.APPLICATION_MODAL);
-                stage.initOwner(node);
-                stage.setOnHidden((stageEvent) -> {
-                    indexD.refresh();
-                });
-                mdc.showModal(stage);
-
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
+            if (id != null)
+                displayEdit(event);
+            else
+                displayWarning();
         });
 
         btnDelete.setOnAction((event) -> {
-            dao.delete(id);
-            indexD.getSelectionModel().getSelectedItem().getParent().getChildren().remove(id - 1);
-            indexD.refresh();
-            log.info("Item deleted.");
+            if (id != null)
+                confirmDialog();
+            else
+                displayWarning();
         });
         // TODO add search filter
     }
+
+    /**
+     *
+     * Class Methods
+     *
+     */
 
     static ObservableList<Desparasitaciones> loadTable(ObservableList<Desparasitaciones> desparasitaciones) {
         List<Desparasitaciones> list = dao.displayRecords();
         for (Desparasitaciones item : list)
             desparasitaciones.add(item);
         return desparasitaciones;
+    }
+
+    private void displayEdit(Event event) {
+        Parent rootNode;
+        Stage stage = new Stage();
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/fxml/deworming/modalDialog.fxml"));
+        Window node = ((Node) event.getSource()).getScene().getWindow();
+        try {
+            rootNode = (Parent) fxmlLoader.load();
+            ModalDialogController mdc = fxmlLoader.getController();
+            mdc.setObject(desparacitaciones);
+            stage.setScene(new Scene(rootNode));
+            stage.setTitle("Desparasitaciones");
+            stage.initModality(Modality.APPLICATION_MODAL);
+            stage.initOwner(node);
+            stage.setOnHidden((stageEvent) -> {
+                indexD.refresh();
+            });
+            mdc.showModal(stage);
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private void confirmDialog() {
+        Alert alert = new Alert(AlertType.CONFIRMATION);
+        alert.setTitle("Confirmación");
+        alert.setHeaderText("Confirmar acción.");
+        alert.setContentText("¿Desea eliminar el registro?");
+
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == ButtonType.OK) {
+            dao.delete(id);
+            indexD.getSelectionModel().getSelectedItem().getParent().getChildren().remove(id - 1);
+            indexD.refresh();
+            log.info("Item deleted.");
+        }
+    }
+
+    private void displayWarning() {
+        Alert alert = new Alert(AlertType.WARNING);
+        alert.setTitle("Advertencia.");
+        alert.setHeaderText("Elemento vacío.");
+        alert.setContentText("No se seleccionó ningún elemento de la lista. Elija un ítem e intente nuevamente.");
+
+        alert.showAndWait();
     }
 }
